Moving CoursewareRedirect into courseware
