+++ src/courseware/index.js
@@ -1,1 +1,2 @@
 export { default } from './CoursewareContainer';
+export { default as CoursewareRedirect } from './CoursewareRedirect';
